requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
gunicorn==20.1.0
Pillow==9.2.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-barcode==0.14.0
sqlparse==0.4.2
